requirements.txt
flask==2.0.1
werkzeug==2.0.1
requests>=2.31.0,<3.0.0
python-dotenv==0.19.0
openai==0.27.0
anthropic==0.3.0
google-generativeai==0.3.0
mistralai==0.0.7
cohere==4.0.0
gunicorn==20.1.0 